Move unknown directive tests into tests directory
